Feat: 창고 상세 페이지 데이터 조회

+++ src/main/java/com/ssg/secondproject/controller/web/WarehouseController.java
@@ -16,7 +16,7 @@ public class WarehouseController {
     }
 
     @GetMapping("/detail")
-    public String warehouseDetail(@RequestParam("id") Long id, Model model) {
+    public String warehouseDetail(@RequestParam("id") Integer id, Model model) {
         model.addAttribute("id", id);
         return "warehouse/detail";
     }
